backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pdfplumber==0.10.4
sqlalchemy==2.0.27
pymysql==1.1.0
python-dotenv==1.0.1
openai==1.12.0
httpx==0.26.0
